feat & style: navbar added to Kitchen and Order
