--- src/Components/Kitchen/modal.tsx
@@ -1,1 +1,0 @@
-
